Move writing of a style part to a dedicated writer. Better encapsulation, more modular.
It was necessary to provide acces to Load* methods from both loading and saving, so OpenXmlHelper for now.
